backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Olivier Risser-Maroix <orissermaroix@gmail.com>", "Mohamed Chelali <mouha.chelali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
requests = "^2.32.2"
numpy = "^1.26.4"
weaviate-client = "^4.6.3"
boto3 = "^1.34.115"
pydantic-settings = "^2.2.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
